Fix usage of tolist in graph_classification/util.py
Currently the way tolist is used, it will not yield an array
if the train and test files only have 1 index in them. This
causes python to throw a type error saying that you can't index
a scalar value.
This will most likely not happen in practice but is still a bug.